+++ src/types/elements/elements.ts
@@ -4,11 +4,20 @@ import type {
   Subscription,
   TextElementEvents,
   CardElementEvents,
+  CardNumberElementEvents,
+  CardExpirationDateElementEvents,
+  CardVerificationCodeElementEvents,
 } from './events';
 import type {
   CreateCardElementOptions,
+  CreateCardExpirationDateElementOptions,
+  CreateCardVerificationCodeElementOptions,
   CreateTextElementOptions,
   UpdateCardElementOptions,
+  CreateCardNumberElementOptions,
+  UpdateCardNumberElementOptions,
+  UpdateCardExpirationDateElementOptions,
+  UpdateCardVerificationCodeElementOptions,
   UpdateTextElementOptions,
 } from './options';
 import type { AtomicBanks, AtomicCards, Tokenize, Tokens } from './services';
@@ -29,6 +38,38 @@ type CardElement = BaseElement<UpdateCardElementOptions, CardElementEvents>;
 
 type TextElement = BaseElement<UpdateTextElementOptions, TextElementEvents>;
 
+type CardNumberElement = BaseElement<
+  UpdateCardNumberElementOptions,
+  CardNumberElementEvents
+>;
+
+type CardExpirationDateElement = BaseElement<
+  UpdateCardExpirationDateElementOptions,
+  CardExpirationDateElementEvents
+> & {
+  month(): ElementWrapper<CardExpirationDateElement>;
+  year(): ElementWrapper<CardExpirationDateElement>;
+};
+
+type CardVerificationCodeElement = BaseElement<
+  UpdateCardVerificationCodeElementOptions,
+  CardVerificationCodeElementEvents
+>;
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type ElementWrapper<T extends BaseElement<any, any> = BaseElement<any, any>> = {
+  element: T;
+  method?: string;
+};
+
+type ElementValue =
+  | TextElement
+  | CardElement
+  | CardNumberElement
+  | CardExpirationDateElement
+  | CardVerificationCodeElement
+  | ElementWrapper;
+
 interface BasisTheoryElements extends Tokenize {
   /**
    * @deprecated use {@link tokens} or {@link tokenize}
@@ -42,6 +83,18 @@ interface BasisTheoryElements extends Tokenize {
 
   createElement(type: 'card', options?: CreateCardElementOptions): CardElement;
   createElement(type: 'text', options: CreateTextElementOptions): TextElement;
+  createElement(
+    type: 'cardNumber',
+    options: CreateCardNumberElementOptions
+  ): CardNumberElement;
+  createElement(
+    type: 'cardExpirationDate',
+    options: CreateCardExpirationDateElementOptions
+  ): CardExpirationDateElement;
+  createElement(
+    type: 'cardVerificationCode',
+    options: CreateCardVerificationCodeElementOptions
+  ): CardVerificationCodeElement;
 }
 
 interface BasisTheoryElementsInternal extends BasisTheoryElements {
@@ -62,6 +115,11 @@ export type {
   BaseElement,
   CardElement,
   TextElement,
+  CardNumberElement,
+  CardExpirationDateElement,
+  CardVerificationCodeElement,
+  ElementWrapper,
+  ElementValue,
   BasisTheoryElements,
   BasisTheoryElementsInternal,
 };