+++ src/types/elements/events.ts
@@ -37,6 +37,12 @@ type TextElementEvents = BaseElementEvents | ChangeEvent;
 
 type CardElementEvents = BaseElementEvents | CardChangeEvent;
 
+type CardNumberElementEvents = BaseElementEvents | CardChangeEvent;
+
+type CardExpirationDateElementEvents = BaseElementEvents | ChangeEvent;
+
+type CardVerificationCodeElementEvents = BaseElementEvents | ChangeEvent;
+
 /**
  * Utility type that helps find a Union type based on a `type` property
  */
@@ -62,6 +68,9 @@ export type {
   BaseElementEvents,
   TextElementEvents,
   CardElementEvents,
+  CardNumberElementEvents,
+  CardExpirationDateElementEvents,
+  CardVerificationCodeElementEvents,
   ElementEventListener,
   Subscription,
 };